backend/requirements.txt
fastapi
uvicorn[standard]
redis[async]
python-jose
passlib
bcrypt==3.2.2
sqlalchemy
asyncpg
fastapi[all]
gym
gymnasium
torch
chess
greenlet